+++ packages/extensions/package.json
@@ -1,12 +1,10 @@
 {
     "name": "extensions",
     "private": true,
-    "version": "0.0.0",
+    "version": "3.0.0",
     "scripts": {
-        "start": "cross-env NODE_ENV=development webpack --watch",
-        "analyzer": "cross-env NODE_ENV=production webpack --hide-modules --config ./webpack.config.analyzer.js",
-        "build": "yarn webpack:build && yarn zip",
-        "webpack:build": "cross-env NODE_ENV=production webpack --hide-modules --config ./webpack.config.js",
+        "start": "cross-env node ./scripts/start.js",
+        "build": "cross-env NODE_ENV=production node ./scripts/start.js",
         "zip": "bestzip chrome.zip dist/*"
     },
     "dependencies": {
