+++ packages/san-devtools/package.json
@@ -5,8 +5,8 @@
     "description": "Development tools for debugging San.",
     "main": "dist/backend.js",
     "scripts": {
-        "start": "node ./start.js",
-        "build": "cross-env NODE_ENV=production webpack --mode production --progress --hide-modules",
+        "start": "cross-env node ./scripts/start.js",
+        "build": "cross-env NODE_ENV=production node ./scripts/start.js",
         "prepublishOnly": "npm run build"
     },
     "repository": {
